requirements.txt
llama-cpp-python
farm-haystack[weaviate]
haystack-ai
sentence_transformers
pypdf
python-box